refactor to more standard project structure
* move build targets that produce binaries into subdirs in `cmd/`
  * `seedtest.go` -> `cmd/seedtest/main.go`
  * `main.go` -> `cmd/server/main.go`
  * moved accompanying files for server to the server dir in cmd
* created `internal/` dir for go code that should only be used internally by this project
* moved static data files into `internal/data/` dir
* removed tmp.txt this must have been accidentally added)
* removed `db000` file
* removed `wetorrent` file (no references to it)
* moved `w64system/` into `internal/`
* adding gitignore
* adding `scripts/` dir
* added `scripts/debian-toolchain-setup.sh` for installing dependencies
  * this may be incomplete, merely what i needed to install on my machine ...
