+++ internal/Webapp/core/core.html
@@ -2,7 +2,7 @@
 <html>
 <head>
 	<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
-	<link href="./styles.css" rel="stylesheet">
+	<link href="styles.css" rel="stylesheet">
 	<link rel="icon" type="image/x-icon" href="./favicon.ico">
 </head>
 <body>
