+++ internal/data/Icons/bundled.txt
