+++ scripts/debian-toolchain-setup.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+apt-get install libglfw3-dev libxcursor-dev libxinerama-dev libxi-dev libgl1-mesa-dev xorg-dev
